# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Training analytics",
    "description": "API over the training services; the DTOs are the same ones the Streamlit UI renders.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/plan/today": {
      "get": {
        "tags": [
          "plan"
        ],
        "summary": "Today's planned workout; decided on the first call, the same plan afterwards",
        "operationId": "get_plan_today",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyDecisionDTO"
                }
              }
            }
          },
          "422": {
            "description": "Invalid or unacceptable parameters / body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/plan/today/regenerate": {
      "post": {
        "tags": [
          "plan"
        ],
        "summary": "Decide today again, optionally for another sport (run | bike); a done workout stays",
        "operationId": "regenerate_today",
        "parameters": [
          {
            "name": "sport",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sport"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyDecisionDTO"
                }
              }
            }
          },
          "422": {
            "description": "Invalid or unacceptable parameters / body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/plan/week": {
      "get": {
        "tags": [
          "plan"
        ],
        "summary": "The ISO week containing `date` (default today): targets, planned vs done per day",
        "operationId": "get_plan_week",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeekPlanDTO"
                }
              }
            }
          },
          "422": {
            "description": "Invalid or unacceptable parameters / body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/plan/goal": {
      "get": {
        "tags": [
          "plan"
        ],
        "summary": "The active goal, or null",
        "operationId": "get_plan_goal",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/GoalDTO"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Plan Goal"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "plan"
        ],
        "summary": "Set the goal race (replaces the active goal)",
        "operationId": "put_plan_goal",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalDTO"
                }
              }
            }
          },
          "422": {
            "description": "Invalid or unacceptable parameters / body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "plan"
        ],
        "summary": "Clear the goal (the season falls back to the no-goal cycle)",
        "operationId": "delete_plan_goal",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        }
      }
    },
    "/api/plan/{planned_id}": {
      "get": {
        "tags": [
          "plan"
        ],
        "summary": "One planned workout with its steps",
        "operationId": "get_planned",
        "parameters": [
          {
            "name": "planned_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Planned Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedWorkoutDTO"
                }
              }
            }
          },
          "404": {
            "description": "The requested entity does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Invalid or unacceptable parameters / body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/plan/{planned_id}/status": {
      "post": {
        "tags": [
          "plan"
        ],
        "summary": "Mark a planned workout done, skipped, or back to planned",
        "operationId": "post_plan_status",
        "parameters": [
          {
            "name": "planned_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Planned Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedWorkoutDTO"
                }
              }
            }
          },
          "404": {
            "description": "The requested entity does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Invalid or unacceptable parameters / body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DailyDecisionDTO": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "workout": {
            "$ref": "#/components/schemas/PlannedWorkoutDTO"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "rule": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule",
            "description": "§10.4 rule that decided (1–4); null when an existing plan was kept"
          },
          "readiness": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Readiness",
            "description": "persisted readiness of the day, 0–100"
          },
          "readiness_band": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Readiness Band",
            "description": "\"green\" | \"yellow\" | \"red\""
          },
          "acwr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acwr",
            "description": "ACWR[D−1], the last complete day (rule input)"
          },
          "tsb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tsb",
            "description": "TSB[D] (rule input)"
          },
          "created": {
            "type": "boolean",
            "title": "Created",
            "description": "true: decided just now; false: an existing plan was returned"
          }
        },
        "type": "object",
        "required": [
          "date",
          "workout",
          "reason",
          "rule",
          "readiness",
          "readiness_band",
          "acwr",
          "tsb",
          "created"
        ],
        "title": "DailyDecisionDTO",
        "description": "GET /plan/today – the day's planned workout with the inputs the decision used (METRICS §10.4)."
      },
...
      "ErrorBody": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorBody"
      },
      "GoalDTO": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "race_date": {
            "type": "string",
            "format": "date",
            "title": "Race Date"
          },
          "distance_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance M",
            "description": "metres"
          },
          "target_time_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Time S",
            "description": "seconds"
          },
          "sport": {
            "type": "string",
            "title": "Sport",
            "description": "\"run\" | \"bike\""
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "race_date",
          "distance_m",
          "target_time_s",
          "sport",
          "active"
        ],
        "title": "GoalDTO"
      },
      "GoalIn": {
        "properties": {
          "race_date": {
            "type": "string",
            "format": "date",
            "title": "Race Date",
            "description": "not in the past"
          },
          "distance_m": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance M",
            "description": "metres, e.g. 10000"
          },
          "target_time_s": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Time S",
            "description": "seconds"
          },
          "sport": {
            "type": "string",
            "title": "Sport",
            "description": "\"run\" | \"bike\"",
            "default": "run"
          },
          "active": {
            "type": "boolean",
            "title": "Active",
            "description": "false stores the goal without using it (all others off)",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "race_date"
        ],
        "title": "GoalIn",
        "description": "PUT /plan/goal body: the race the season is built towards. Replaces the active goal."
      },
...
      "PlannedWorkoutDTO": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "sport": {
            "type": "string",
            "title": "Sport",
            "description": "\"run\" | \"bike\" | \"rest\""
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "key": {
            "type": "string",
            "title": "Key",
            "description": "§10.7 library key, empty when unknown"
          },
          "slot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot",
            "description": "weekday role filled: \"easy\" | \"long\" | \"q1\" | \"q2\""
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "\"planned\" | \"pushed\" | \"done\" | \"skipped\""
          },
          "missed": {
            "type": "boolean",
            "title": "Missed",
            "description": "planned/pushed and its date is before today"
          },
          "estimated_load": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Load",
            "description": "§10.6, TSS-equivalent points"
          },
          "duration_s": {
            "type": "integer",
            "title": "Duration S",
            "description": "total planned seconds, repeats multiplied"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason",
            "description": "one-line Slovak reason of the decision"
          },
          "completed_activity_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed Activity Id"
          },
          "actual_load": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Load",
            "description": "load of the completed activity (its primary load)"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/WorkoutStepDTO"
            },
            "type": "array",
            "title": "Steps"
          },
          "structure": {
            "additionalProperties": true,
            "type": "object",
            "title": "Structure",
            "description": "the raw §10.5 JSON"
          }
        },
        "type": "object",
        "required": [
          "id",
          "date",
          "sport",
          "name",
          "key",
          "slot",
          "status",
          "missed",
          "estimated_load",
          "duration_s",
          "reason",
          "completed_activity_id",
          "actual_load",
          "steps",
          "structure"
        ],
        "title": "PlannedWorkoutDTO",
        "description": "A planned workout (METRICS §10.5); `sport = \"rest\"` is a rest day with no steps."
      },
...
      "StatusIn": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "\"done\" | \"skipped\" | \"planned\" (undo)"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "StatusIn",
        "description": "POST /plan/{id}/status body."
      },
...
      "WeekDayDTO": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "weekday": {
            "type": "string",
            "title": "Weekday",
            "description": "\"mon\" … \"sun\""
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "template role: \"rest\" | \"easy\" | \"long\" | \"q1\" | \"q2\""
          },
          "planned": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlannedWorkoutDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "actual_load": {
            "type": "number",
            "title": "Actual Load",
            "description": "load_total of the day (all activities)"
          },
          "activities": {
            "type": "integer",
            "title": "Activities",
            "description": "number of activities on the day"
          }
        },
        "type": "object",
        "required": [
          "date",
          "weekday",
          "role",
          "planned",
          "actual_load",
          "activities"
        ],
        "title": "WeekDayDTO"
      },
      "WeekPlanDTO": {
        "properties": {
          "monday": {
            "type": "string",
            "format": "date",
            "title": "Monday"
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "recovery": {
            "type": "boolean",
            "title": "Recovery"
          },
          "target_load": {
            "type": "number",
            "title": "Target Load"
          },
          "run_target": {
            "type": "number",
            "title": "Run Target"
          },
          "bike_target": {
            "type": "number",
            "title": "Bike Target"
          },
          "done_load": {
            "type": "number",
            "title": "Done Load",
            "description": "Σ load_total of the week's days"
          },
          "remaining_load": {
            "type": "number",
            "title": "Remaining Load",
            "description": "max(0, target_load − done_load)"
          },
          "days": {
            "items": {
              "$ref": "#/components/schemas/WeekDayDTO"
            },
            "type": "array",
            "title": "Days",
            "description": "Monday … Sunday"
          }
        },
        "type": "object",
        "required": [
          "monday",
          "phase",
          "recovery",
          "target_load",
          "run_target",
          "bike_target",
          "done_load",
          "remaining_load",
          "days"
        ],
        "title": "WeekPlanDTO",
        "description": "GET /plan/week – the ISO week containing the requested date: targets, planned vs done per day."
      },
...
      "WorkoutStepDTO": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "\"warmup\" | \"work\" | \"recovery\" | \"cooldown\" | \"steady\""
          },
          "duration_s": {
            "type": "integer",
            "title": "Duration S",
            "description": "seconds of one round (not multiplied by `repeat`)"
          },
          "target_kind": {
            "type": "string",
            "title": "Target Kind",
            "description": "\"hr_zone\" | \"pace_range\" | \"open\""
          },
          "zone": {
            "type": "integer",
            "title": "Zone",
            "description": "1–5: HR zone, pace zone, or the effort zone of an open step"
          },
          "low_mps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Low Mps",
            "description": "pace_range lower speed bound, m/s"
          },
          "high_mps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "High Mps",
            "description": "pace_range upper speed bound, m/s"
          },
          "repeat": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repeat",
            "description": "repeat count when the step is inside a repeat, else null"
          },
          "group": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group",
            "description": "index of the repeat block (0, 1, …), else null"
          }
        },
        "type": "object",
        "required": [
          "type",
          "duration_s",
          "target_kind",
          "zone",
          "repeat",
          "group"
        ],
        "title": "WorkoutStepDTO",
        "description": "One plain step of a workout, flattened for display; steps of a repeat share `group` and `repeat`."
      },
...
    }
  }
}